Workaround MSVC-2017 bugs after Chromium 69-based merge

+++ chromium/device/fido/public_key_credential_params.cc
@@ -35,7 +35,8 @@ PublicKeyCredentialParams::CreateFromCBORValue(
     }
 
     credential_params.push_back(PublicKeyCredentialParams::CredentialInfo{
-        CredentialType::kPublicKey, algorithm_type_it->second.GetInteger()});
+        CredentialType::kPublicKey,
+        static_cast<int>(algorithm_type_it->second.GetInteger())});
   }
 
   return PublicKeyCredentialParams(std::move(credential_params));
